Allow configuring certain settings per-controller type
Adds a ControllerSettings resource. Defaults are provided for the
Vive wands, and can be provided for any controller. There is a
separate config file to allow the user to set or override any of
them.

The settings define the offset of the hands, and whether a boolean
action needs to be pressed to turn or walk. These boolean actions
can be useful for trackpads, but disabled for sticks, for example.

The demo room now allows saving settings manually by pressing space
or enter, as a temporary workaround for godotengine/godot#70755.