Reimplement snap turning
This was the turning behavior in the 3.x version.
It is now enabled by default. The previous smooth turning behavior
can be switched to in the settings. The smooth turn also now only
turns the player's body, rather than the root node, fixing issues
with the hands not moving while turning.