Update Readme
Add a note about Vulkan pipelines and the status of asynchronous
shader and pipeline compilation in Godot 4. Recommend 4.1 or later
because of it.

Fix a few spelling mistakes and tweak some wordings.

Add line breaks within paragraphs. Each line is now under 100
characters. It should be readable with word wrap off now.